requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-environ==0.11.2
Pillow==10.1.0
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
drf-spectacular==0.26.5
django-phonenumber-field==7.2.0
phonenumbers==8.13.25
django-storages==1.14.2
boto3==1.29.7
django-extensions==3.2.3
